🧬 Extract the base data modals into new module *_base* for common usage in sub-package *openapi*.
